// src/models/filters.model.ts
export class Filters {

  //Fields
  // ageRange:any={
  //   upper:35,
  //   lower:25
  // };

  ageRangeUpper: number;
  ageRangeLower: number;
  maxDist: number;
  female: boolean;
  male: boolean;
  meetingType: string;
  addictsType: string;
}

  // constructor (ageRangeUpper: number,
  //              ageRangeLower: number,
  //              maxDist: number,
  //              female: boolean,
  //              male: boolean,
  //              meetingType: string,
  //              addictsType: string){
  //
  //   this.ageRangeUpper = ageRangeUpper;
  //   this.ageRangeLower = ageRangeLower;
  //   this.maxDist = maxDist;
  //   this.female = female;
  //   this.male = male;
  //   this.meetingType = meetingType;
  //   this.addictsType = addictsType;
  // }}